Split sys module in test-runner into one per platform
